Fix test namespaces #72
