+++ tests/GitHubRepoTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace tests\Devboard\GitHub\Fetch\Repo;
+namespace tests\Devboard\GitHub;
 
 use Devboard\GitHub\GitHubRepo;
 use Devboard\GitHub\Repo\GitHubRepoApiUrl;
